+++ test/gpn_test.exs
@@ -1,0 +1,12 @@
+defmodule SimpleCardBrandGPN do
+  @moduledoc """
+  GPN tests.
+  """
+  use ExUnit.Case
+
+  test "GpnWikipedia" do
+    for first <- ["1", "2", "6", "7", "8", "9"] do
+      assert SimpleCardBrand.card_brand("1", 16) == {:ok, :gpn}
+    end
+  end
+end